fix the ohos app development scenario, SIGILL stack incompleted, enable multi process dump, enable frame vector output, restruct log util.

+++ tools/process_dump/dfx_util.cpp
@@ -83,7 +83,7 @@ bool TrimAndDupStr(const std::string &source, std::string &str)
     }
 
     uint32_t maxStrLen = NAME_LEN;
-    uint32_t offset = end - begin;
+    uint32_t offset = (uint32_t)(end - begin);
     if (maxStrLen > offset) {
         maxStrLen = offset;
     }
